Adds a required_flags field to Arcs build flag definitions.
This lets you list other flags that your feature relies upon. These flags must be enabled in order for your flag to be enabled.

This CL adds copious error checking at build time. It ensures that all flag overrides respect the required_flag dependencies (i.e. if you enable a flag then any flags it requires are also enabled), and also checks that the definitions themselves are sane (i.e. if you mark a flag as READY, then none of the flags it requires are allowed to be NOT_READY).

In a follow-up CL I will add runtime checks as well, for unit testing.

PiperOrigin-RevId: 344466848