Converts the generated BuildFlags class to Kotlin.
Not just for fun; in a follow-up CL I'm going to switch the var fields to use getters and setters instead, and then hook into the setter methods to validate that the flags are all set correctly.

Moved the invokation of the kt_library inside the generate_build_flags macro to make it easier to use. Still have to wrap the kt_library inside a java_library so that I can use neverlink (kt_library doesn't support it).

PiperOrigin-RevId: 345142169