Makes it painfully obvious where the global StorageKeyManager is being used.
Replaces the StorageKeyParser companion object (which is actually a StorageKeyManager instance, not a StorageKeyParser instance) with a GLOBAL_INSTANCE field on StorageKeyManager, and removes the old companion object. This will make it really easy to find everywhere where the global instance is being used, so that they can slowly be migrated to a dependency passing approach.

PiperOrigin-RevId: 345585107