Simplify VolatileDriverProvider impl, remove on close
1) Remove the two-layer provider approach for VolatileDriverProvider.

Insteaad, the VolatileDriverProvier now maintains a map of
VolatileMemoryEntries, and creates new VolatileMemoryImpls directly using those
memories.

This means that we no longer have per-arc remove/count functionality available.
This didn't seem to be used anywhere, and so it appears that this won't be an
issue.

2) Clean up the arcMemories entry when the final provider using it is closed.

This implicit cleanup signal means that today, stopping an arc will result in the
memory for it being deleted. In the event that we need some arc "pausing"
behavior that does not result in volatile memory deletion, we should implement
pause in such a way that it provides a signal to the storage layer to not clean
up yet. This can be as simple as not closing the EntityHandleManager in the
ArcHost, so that its storage channels stay open.

3) Add a test to verify that drivers for the same storage key use the same
memory (and do a bit of cleanup in the test file).

PiperOrigin-RevId: 345702658