Make DefaultStorageKeyManager private.
Currently DefaultStorageKeyManager is a singleton object, and is aliased as StorageKeyParser's companion object. Everything always uses it via that companion object, except a single test that uses DefaultStorageKeyManager directly. I've updated that usage to be consistent with everything else, and made it private. This is the first baby step on the journey to make it no longer a global static singleton.

PiperOrigin-RevId: 345983827