Defines a DevModeBuildFlags base class that contains all the complex flag logic.
This class is only to be used for dev-mode BuildFlags (i.e. in unit testing, or in test apps). It dramatically simplifies the build flags code generator, by pulling common code into a base class. It also allows much easier testing of the flag logic itself.

Will be useful for parameterized testing of build flags in the future, since it's now possible to dynamically update flags just by calling the update() method.

PiperOrigin-RevId: 345983975