Ignore model updates received when in READY_TO_SYNC.
Most of the time the StorageProxy moves from READY_TO_SYNC to AWAITING_SYNC, sends a sync request then moves to SYNC upon receiving the subsequent model update. Rarely, a different actor may cause the backing storage to send a model update to all connected proxies, while one of them is in the READY_TO_SYNC state (i.e. hasn't yet sent a sync request). This is not actually an error. It would be possible to use that update as the sync (and don't send a request), but that needs to be deferred until the correct time, so it's easier to just ignore it.

PiperOrigin-RevId: 345989915