Renames DefaultStorageKeyManager to StorageKeyManagerImpl, moves it to its own file, and makes it a proper class instead of a singleton object.
This will let us reuse it in tests without relying on it being a global singleton. Also creates DummyStorageKeyManager for use in tests.

It won't let you change the parsers it knows about. In fact, arguably the reset and addParser methods shouldn't even be on StorageKeyManager, but that's a problem for another day.

PiperOrigin-RevId: 346211885