Don't call idle immediately after closing StorageEndpoint
Unbinding from an Android service is only an indication that the client no
longer wants connected/disconnected events from the storage service, and that
from the client's point of view, it is OK for the service to shut down.

But it *doesn't* make any acquired bindings stop work. This means that we
called close, and then called idle, which succeeds, since the binding still
works.

The `idle` call coming in after `onUnbind` was causing a store to get
re-created (and then never cleaned up, since the binding in question would
never be bound/unbound again).

PiperOrigin-RevId: 346218079