Uses volatile driver in transient Arcs.
The @ttl('30m') annotation used to force the intermediate data to the in-memory database driver, where it would be cleaned up after the TTL and GC jobs run.

The downside of @ttl('30m') was that data would only get cleaned up after 2+days. Now after cl/345702658 and cl/346218079 Volatile drivers for an arc are cleaned up after stopping an arc, which provides a more memory efficient way to store intermediate data.

PiperOrigin-RevId: 346811085