Don't include the stringified value by default.
`ReferencablePrimitive`s carry around a "valueRepr" field, introduced to make handling ByteArrays better, by default. This is not needed most of the time, but adds to the memory overhead of all primitive values in raw entities.

This change minimizes storing valueRepr (pretty much, only for the ByteArray use case) and recomputes the string when needed.

Overhead savings still needs to be calculated.