Move BindingContextStats to a general-purpose utility package
This is some re-organization in preparation of adding some unit tests for these
classes.

Nothing about the classes were specific to the storage service's
BindingContext. This CL moves the relevant files to a package in
arcs.core.util, renames the to something more general, and splits testable
classes into their own files.

There are no functional or structural changes to the classes themselves, only
moves and renames.

Unit tests will be added in a follow-up CL

PiperOrigin-RevId: 347856786