Refactor StorageCommunicationEndpointProvider to return StorageCommunicationEndpoint per StoreInfo (#6614)
This patch moves the StoreInfo argument from the
getStorageEndpointProvider() call to the getStorageEndpoint() call, which means that
a single StorageEndpointProvider can be used for a set of StorageProxies.

Eventually this single StorageEndpointProvider will become a separate
StorageEndpointManager, rather than being a responsibility of Stores directly.
This will allow the same manager to be used in both direct communication mode
and in hosted (via PEC) mode.

A short-term consequence is that both StoreInfo and an ActiveStore are passed
into the StorageProxy constructor (the ActiveStore is currently still the
StorageEndpointProvider and StoreInfo is needed to get the endpoint). This
duplication will be resolved in a subsequent patch (the ActiveStore argument
will become a StorageEndpoint instead, vended by a StorageEndpointManager
which will have responsibility for creating ActiveStores on demand).
________________________________________________________________________________________________

next steps:
- DirectStorageEndpointManager should implement StorageCommunicationEndpointProvider, and ActiveStore StorageCommunicationEndpoint instead.

Closes #6614

PiperOrigin-RevId: 347933211