Don't block the ReferenceModeStore receive queue while waiting for pendingIds to resolve.
Some time ago, a timeout-based solution was implemented to prevent a situation
where a ReferenceModeStore will hang indefinitely when it's impossible to
resolve a reference stored int he container store. In the particular case in
question, this was due to partially-written database data existing after an app
crashing or being killed: the container items had been persisted, but the
backing store items had not all been fully written.

(Original bug: b/158490596)
(Original fix: https://github.com/PolymerLabs/arcs/pull/5521)

The implementation implements a timeout by waiting for a deferred to be
resolved inside of a `withTimeout` block. However, the `deferred.await` runs
synchronously with respect to the receive queue, preventing the receive queue
from progressing, which means that it's not actually capable to reciving and
processing messages from the backing store that would resolve these pending IDs.

In this change, we resolve the issue by launching a different coroutine to
manage the timeout for a sync request.

Additionally, a new test is added to verify that in cases where the backing
data does eventually arrive, a timeout is not hit.

PiperOrigin-RevId: 351249979