Add tests for CrdtEntity.merge() (#6810)
Add tests in line for the test plan of CrdtEntity's merge function. This is the first in a series of PRs around this class. This PR focuses on various combinations of merges and operations along with singletons in a merge.

This PR also includes an added test for CrdtSet to highlight an underlying issue discovered.

Question for @shans: In the `crdtEntity_merge_singletonFieldTypeMismatch` test, we have mismatched field types for the singleton "foo". In this case the "foo" field becomes null after a merge. Is this correct?

Closes #6810

COPYBARA_INTEGRATE_REVIEW=https://github.com/polymerlabs/arcs/pull/6810 from SHeimlich:CrdtEntityTests 1ed8d85ace55d7f86fe4bb3f27094ddcd89c2af0
PiperOrigin-RevId: 351401183