Silence type errors reported by TS 4.1 for Promise resolve() functions
This change is needed to unblock upgrading google3 to TypeScript 4.1.
It is part of an LSC documented at go/ts41-promise-errors-lsc

TS 4.1 improves the type checking of Promise and related types.
In particular it now requires that the `resolve` method that gets passed
to the function you give the `Promise<T>` constructor must take an argument
of a type that matches the type `T`.

See also b/171742909 and go/ts41upgrade

This change only silences the error by casting the problem value to a custom
alias for `any` named `AnyDuringTs41Migration`. This is guaranteed to leave
the behavior of the code unchanged.

If you see how to really fix this code and eliminate the cast, please do that
in a followup CL and test it against the upgrade CL listed at the top of
go/ts41upgrade to confirm that you've really fixed it.

NOTE: Just removing the cast won't trigger any errors while google3 is still
on TS 4.0, but it will require us to make this change again before we
can upgrade to TS 4.1.

Please do not ask the author of this CL to make any changes to it unless you
believe your code will actually be broken by this change. We need to change
hundreds of files, and cannot realistically give them individual attention.

Tested:
    TAP --sample ran all affected tests and none failed
    http://test/OCL:348490377:BASE:348457395:1608576060046:8687b086
PiperOrigin-RevId: 352683448