LSC - Add additional parenthesis around assignment within conditionals
Assignment within parenthesis is both-
1. A common mistake made by developers intending to check equality that can be detected by linters
2. A valid pattern that we still want to support

By adding the additional set of parenthesis for intended cases we are able to apply the lint rule to find mistakes without false positives.

More details in the LSC Doc - go/lsc-ts-assign-in-parens

When reviewing, please verify that this usage of assignment in parenthesis is intentional and not a bug.

Tested:
    TAP --sample ran all affected tests and none failed
    http://test/OCL:354208792:BASE:354158574:1611800845321:1f2a62c4
PiperOrigin-RevId: 354355375