Replaces a variety of different removeExpiredEntities_* test cases in DatabaseImplTest with two tests (one for singletons and one for collections) that use FixtureEntity, so that all field types are comprehensively tested.
I added a bunch of logic to FixtureEntities to compute db-related constants for use in the test. This is fairly verbose, but should make it more robust to changes to the FixtureEntity schema in the future.

Implements some of the recommendations from the testing plan: https://docs.google.com/document/d/1ulqzCM8kMfbwrtJjn6I5R2vO49nXY9Dtp6iTNLU41mQ/edit#

PiperOrigin-RevId: 354412352