Add unit tests for Allocator
Adds unit tests for Allocator. This is done in a new class: AllocatorUnitTest.kt, rather than in AllocatorTestBase because the latter is more akin to an integration test. The tests in this CL are limited in their reach to allow full coverage of Allocator.

Also - fixes a deadlock when stopArc tried to re-lock the mutex which already had a lock held by startArcForPlan.

PiperOrigin-RevId: 355326696