[low precision] Duration support for Kotlin (#6881)
Added both Shane and Cameron with the intent of dividing the review workload

Shane for review of Typescript changes
Cameron for review of Kotlin changes

Note: @shans I think this is all the different places that currently need to be updated to add a type. I'm planning to come back to this to try to reduce the workload (per type added).

Closes #6881

PiperOrigin-RevId: 355328215