Factor out common test helper code into ReferenceModeStoreTestBase.
This logic is now shared between all of ReferenceModeStoreTestBase, ReferenceModeStoreDatabaseIntegrationTest and ReferenceModeStoreDatabaseImplIntegrationTest. Almost all of the helper methods and constants have been deduped and moved into the base class.

The major change made to the tests is that they now all have a list and an inline entity field (previously only the DatabaseImpl variant had those).

In a follow up CL I will start pulling the identical tests into the base class too.

PiperOrigin-RevId: 355723645