Replace the stores ConcurrentHashMap in the `StorageService` with the new `ReferencedStores` data-structure. This involved modifying `StorageServiceManager` and `DevToolsStorageManager` to use the `ReferencedStores` data-structure as well.
Replace the internal cache of stores in `StorageServiceNgImpl` with a provided `ReferencedStores` data-structure. This removes the need for the `ChannelCounter` class.

Modified `StorageChannelImpl` to take a `ReleasableStore`. This replaces the need for `onCloseCallback`.

`ReferencedStores` has an additional method to be compatible with `BindingContext` (`remove`).

`ReferencedStores` has an additional method to be used only for testing (`createReleasableStore`).

PiperOrigin-RevId: 355745722