Moves most tests from ReferenceModeStoreDatabaseImplIntegrationTest into ReferenceModeStoreTestBase.
These tests have all been updated to use the DriverTestHelper rather than the DatabaseDriver directly, so work for all driver backends, and have been deduped from the other ReferenceModeStore tests.

There's a handful of tests left in the other test files that still need to be updated and moved into the TestBase.

PiperOrigin-RevId: 356167239