Nullable proto parcelable (#6938)
Implements the core of Nullable support in Kotlin (conversion to/from Referencable, to/from Proto and Kotlin, Parcelable and Proto representations of Nullable Types).

Also turns on Nullable Support TypeScript flag (needed to ensure parsing and code generation of Nullable is available for Kotlin testing). This should not have an impact on Kotlin runtime behaviour and so is deemed safe.

Closes #6938

PiperOrigin-RevId: 359628137