Remove bespoke map.assertEquals implementation.
* Replace calls with Truth's assertThat(map).containsExactlyEntriesIn()
* Modify CrdtEntity construction to take a default reference builder that
  either deals correctly with input values or throws an exception.
* Update CrdtEntityTest to correctly build CrdtEntity field values

The second change is required for this CL because the
syncRequest_fromProxy_withPendingIds test was incorrectly wrapping
RawEntity and PrimitiveList fields inside CrdtEntity, then relying
on incorrect value matching logic in map.assertEquals to pass.

The third change is triggered by the second; CrdtEntityTest was incorrectly
using the internal CrdtEntity.ReferenceImpl class directly and this started
causing the new and pickier default reference builder to throw exceptions.

PiperOrigin-RevId: 359639066