Added inline entity and reference field generation to fuzz testing infrastructure.
This makes entity generation recursive; entities can generate entity fields, schemas can generate schema fields, etc. To simplify the process of building things
recursively from generators this patch also adds generatorWithRecursion and transformerWithRecursion functions to the fuzz testing infrastructure. Both functions allow Generator construction to be expressed as a function that accepts
a Generator of the same type; generatorWithRecursion also allows the specification
of a maximum recursion depth and terminal Generator.

This CL also uncovered an issue with deserialization of parcels, which is fixed in
cl/361476803.

PiperOrigin-RevId: 361737266