Catch common cases where Directions are misspelled  (#6996)
Unfortunately, because Direction is optional, we can't simply catch any case where an unknown direction is used (as the unknown text might actually be a handle name or type etc).

I think we're starting to run into the limits of what PEGjs can do for us in terms of error messages, but we can continue to add 'common sense' error cases like this when they are common and confusing enough.

Closes #6996

PiperOrigin-RevId: 363521927