Automatic type slicing of entities for Kotlin particles - step 1.
This CL updates the handle classes to accept separate types for reading and writing entities. The writing type will be the "slice" interface for an entity, but in this CL that has been aliased to refer to the concrete entity class. Thus this CL contains no actual logic changes, apart from the codegen update to produce the placeholder slice aliases.

Step 2 will introduce the logic change behind a flag as appropriate.

PiperOrigin-RevId: 364468238