Update most unit tests to the new type slicing interfaces for Kotlin.
This moves as much of the testing code to the new <Entity, EntitySlice> generic typing as possible. This touches a single non-test file with a change in the generic type param, and as such will be type erased ar runtime.

Therefore this does not affect production and should not change any test behaviour.

There are a few places (e.g. imports or specific type aliases) that can't be changed without updating the codegen to produce the correct type slicing classes, but they can also be left as-is until after the full feature has landed.

PiperOrigin-RevId: 364924674