Update codegen to output correct type slicing interfaces.
Generated entity classes will have a corresponding "slice" interface
that enables automatic type conversion between entities within a
particle.

This is behind a build flag, both for the codegen of the interfaces
themselves and for the new unit test. The latter introduces a build rule
flag for conditional output of build targets based on the Arcs build
flag infrastructure.