Fixes a bug in ManyOfParser where the location of failure in the composed parser was incorrectly reported as the end of a successfully parsed range.
If the parser composed by ManyOfParser is itself composed, the failure location
may not correspond to the end of the last successfull "full parse" of the parser
composed by ManyOfParser.

The fix is to use the end of the last successful parse as the end of the range
successfully parsed by ManyOfParser.

PiperOrigin-RevId: 366171069