Fully implement serialization of short-form storage key protocols.
e.g. Instead of `ramdisk://abc` it is now `r|abc`

Updated lots of tests to be parameterized based on the STORAGE_KEY_REDUCTION flag. Had to introduce a temporary variable to make assertions that hardcode the protocol work with both versions of the flag. Should be easy to clean up again later.

PiperOrigin-RevId: 369137971