Update DatabaseImpl to support the shortened inline storage key protocol.
That is, `i|` instead of `inline://`

Note that these changes are not behind a flag, because the SQL statements were verbose enough already and I didn't want to copy them. The additions should be harmless while the flag is off.

Also updated the test to be parameterized by the STORAGE_KEY_REDUCTION flag, and increased the number of shards from 5 to 20, to account for the massively increased number of tests being run (2 flag parameters => 4x total)

PiperOrigin-RevId: 369144422